feat: macOS renderer module - [5/20]
- add a new module `renderer` to the macos platform
- create a new file `renderer.rs` in the macos platform
- implement a function `build_window_config` in the new `renderer.rs` file
- modify the `build_window_config` function in the `renderer` module to now call the function from the macos platform instead
